+++ apps/hyperdrive-trading/src/ui/markets/PoolRow/LpApyCta.tsx
@@ -5,7 +5,6 @@ import { ReactElement } from "react";
 import { useLpApy } from "src/ui/hyperdrive/lp/hooks/useLpApy";
 import { LpApyStat } from "src/ui/markets/PoolRow/LpApyStat";
 import { PoolStat } from "src/ui/markets/PoolRow/PoolStat";
-import { HyperVueMilesIconUrl } from "src/ui/rewards/HyperVueMilesIconUrl";
 import { RewardsTooltipContent } from "src/ui/rewards/RewardsTooltip/RewardsTooltipContent";
 import { useAccount } from "wagmi";
 
@@ -43,37 +42,32 @@ export function LpApyCta({ hyperdrive }: LpApyCtaProps): ReactElement {
         />
       }
       action={
-        <div className="daisy-indicator w-full sm:w-auto">
-          <span className="daisy-badge daisy-indicator-item p-0">
-            <img src={HyperVueMilesIconUrl} className="size-6 rounded-full" />
-          </span>
-          <Link
-            to="/market/$chainId/$address"
-            params={{
-              address: hyperdrive.address,
-              chainId: hyperdrive.chainId.toString(),
-            }}
-            search={{ position: "lp" }}
-            className="daisy-btn h-10 min-h-10 w-full rounded-full bg-gray-500 sm:daisy-btn-sm hover:bg-gray-400 sm:h-8 sm:bg-gray-600 sm:hover:bg-gray-500 md:w-32"
-            onClick={(e) => {
-              e.stopPropagation();
-              window.plausible("positionCtaClick", {
-                props: {
-                  chainId: hyperdrive.chainId,
-                  poolAddress: hyperdrive.address,
-                  positionType: "lp",
-                  statName: label,
-                  statValue: lpApy?.netLpApy
-                    ? fixed(lpApy.netLpApy).toString()
-                    : "",
-                  connectedWallet: account,
-                },
-              });
-            }}
-          >
-            Add Liquidity
-          </Link>
-        </div>
+        <Link
+          to="/market/$chainId/$address"
+          params={{
+            address: hyperdrive.address,
+            chainId: hyperdrive.chainId.toString(),
+          }}
+          search={{ position: "lp" }}
+          className="daisy-btn h-10 min-h-10 w-full rounded-full bg-gray-500 sm:daisy-btn-sm hover:bg-gray-400 sm:h-8 sm:bg-gray-600 sm:hover:bg-gray-500 md:w-32"
+          onClick={(e) => {
+            e.stopPropagation();
+            window.plausible("positionCtaClick", {
+              props: {
+                chainId: hyperdrive.chainId,
+                poolAddress: hyperdrive.address,
+                positionType: "lp",
+                statName: label,
+                statValue: lpApy?.netLpApy
+                  ? fixed(lpApy.netLpApy).toString()
+                  : "",
+                connectedWallet: account,
+              },
+            });
+          }}
+        >
+          Add Liquidity
+        </Link>
       }
     />
   );