+++ apps/hyperdrive-trading/src/ui/rewards/MilesBanner.tsx
@@ -3,6 +3,7 @@ import { ReactElement } from "react";
 import { useLocalStorage } from "react-use";
 import { ExternalLink } from "src/ui/analytics/ExternalLink";
 import CustomBanner from "src/ui/base/components/CustomBanner";
+import { HyperVueMilesIconUrl } from "src/ui/rewards/HyperVueMilesIconUrl";
 
 export function MilesBanner(): ReactElement | null {
   const { isBannerDismissed, setIsBannerDismissed } = useMilesBannerDismissed();
@@ -12,15 +13,17 @@ export function MilesBanner(): ReactElement | null {
 
   return (
     <CustomBanner
-      className="gradient-hypervue"
+      className="border border-hypervue-red/30 bg-hypervue-pink/20"
       description={
         <>
           <div className="flex w-full justify-between">
             <div className="flex w-full flex-1 items-center justify-center gap-4">
-              <img src="/hyperdrive-solo-logo-white.svg" className="size-8" />
-              <span className="text-base-content">
-                You can now earn Miles by providing liquidity on Hyperdrive!
-                Learn more about the{" "}
+              <img src={HyperVueMilesIconUrl} className="size-8 rounded-full" />
+              <span className="drop-shadow-glow">
+                You can now earn Miles by{" "}
+                <strong className="">providing liquidity</strong> or{" "}
+                <strong className="">opening a short</strong>! Learn more about
+                the{" "}
                 <ExternalLink
                   newTab
                   href="https://mirror.xyz/0xdB081d7cedeDB2cFb4fff2330D9a31f54A025E38/qVENDIYTfUiZw6QZroXBpNxF8UjTGV8YkVq0lOaTznU"