+++ packages/hyperdrive-appconfig/src/base/fixedPoint.ts
@@ -1,5 +1,0 @@
-import * as fixedPoint from "@delvtech/fixed-point-wasm";
-
-fixedPoint.initSync(fixedPoint.wasmBuffer);
-
-export { FixedPoint, fixed, parseFixed } from "@delvtech/fixed-point-wasm";