+++ packages/hyperdrive-appconfig/src/index.ts
@@ -35,7 +35,6 @@ export {
   getAddLiquidityRewardConfigs,
   getOpenShortRewardConfigs,
   getRewardConfig,
-  getYieldSourceRewardConfigs,
 } from "src/rewards/selectors";
 export type {
   AnyReward,