+++ packages/hyperdrive-appconfig/src/rewards/actions/types.ts
@@ -1,8 +1,4 @@
 import { AddLiquidityRewardId } from "src/rewards/actions/lp";
 import { OpenShortRewardId } from "src/rewards/actions/short";
-import { YieldSourceRewardId } from "src/rewards/actions/yieldSource";
 
-export type AnyRewardId =
-  | AddLiquidityRewardId
-  | OpenShortRewardId
-  | YieldSourceRewardId;
+export type AnyRewardId = AddLiquidityRewardId | OpenShortRewardId;