+++ packages/hyperdrive-js/src/base/calculateAprFromPrice.ts
@@ -1,5 +1,5 @@
-import { fixed } from "src/fixed-point";
-import { hyperwasm } from "src/hyperwasm";
+import { fixed } from "@delvtech/fixed-point-wasm";
+import { calcAprGivenFixedPrice } from "@delvtech/hyperdrive-wasm";
 
 /**
  * Calculate the APR of a position given the position duration, the amount paid
@@ -23,7 +23,7 @@ export function calculateAprFromPrice({
    */
   bondAmount: bigint;
 }): bigint {
-  return hyperwasm.calcAprGivenFixedPrice({
+  return calcAprGivenFixedPrice({
     positionDuration,
     price: fixed(baseAmount).div(bondAmount).bigint,
   });