+++ packages/hyperdrive-js/src/hyperwasm.ts
@@ -1,5 +1,3 @@
-// The types resolve, but ESLINT still compalins, so we ignore the error.
-// eslint-disable-next-line import/no-unresolved
 import * as hyperwasm from "@delvtech/hyperdrive-wasm";
 
 export { hyperwasm };